Refactored the code with command handlers and modified README.md and setup.md
