+++ ideas.md
@@ -1,0 +1,16 @@
+# Features
+
+## 1. Leveling for members
+
+1. Users can upvote the user that helped them with the command `/upvote @username keywordformultipoints`. **(Won't be deleted from the channel)**
+   |Keyword|Points|
+   |---|---|
+   | codeError | 2 |
+   | verbalDoubt | 1 |
+   | codeDoubt | 2 |
+   | contribution | 5 |
+   | sharedResource | 2 |
+   | slangUsed | -5 |
+
+1. Level up messages will be sent to the respective user on their DM's
+1. Users can check their information by using `/me` command
